requirements.txt
requests==2.31.0
python-dotenv==1.0.0
pandas==2.1.3
python-dateutil==2.8.2
plotly==5.18.0
numpy==1.26.2

